requirements.txt
fastapi
uvicorn
jinja2
python-multipart
yfinance
pandas
pydantic